+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to oneplace-faucet-batch will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.3] - 2021-04-15
+
+### Added
+- Email Parse Batch
+
+### Fixed
+- Nanopool API now takes shares instead of rating
+
 ## [1.0.2] - 2021-04-12
 
 ### Added